find MS2 using alphatims

+++ alpharaw/thermo.py
@@ -44,7 +44,7 @@ class ThermoRawData(MSData_Base):
             else:
                 masses, intensities = rawfile.GetCentroidMassListFromScanNum(i)
             mz_values.append(masses)
-            intensity_values.append(intensities)
+            intensity_values.append(intensities.astype(np.float32))
             _peak_indices.append(len(masses))
             rt = rawfile.RTFromScanNum(i)
             rt_values.append(rt)
